refactor: contributors and testimonials component and remove commented code
